Updated nvi test data. Red tests
