+++ report-api/src/test/java/no/sikt/nva/data/report/api/fetch/testutils/generator/nvi/TestNviContributor.java
@@ -4,12 +4,11 @@ import java.util.List;
 import no.sikt.nva.data.report.api.fetch.testutils.generator.model.nvi.NviContributorGenerator;
 
 public record TestNviContributor(String id,
-                                 boolean isNviContributor,
-                                 List<TestAffiliation> affiliations) {
+                                 List<NviOrganization> affiliations) {
 
     public NviContributorGenerator toModel() {
         var contributor = new NviContributorGenerator(id);
-        affiliations.stream().map(TestAffiliation::toModel)
+        affiliations.stream().map(NviOrganization::toModel)
             .forEach(contributor::withAffiliation);
         return contributor;
     }
@@ -21,8 +20,7 @@ public record TestNviContributor(String id,
     public static final class Builder {
 
         private String id;
-        private boolean isNviContributor;
-        private List<TestAffiliation> affiliations;
+        private List<NviOrganization> affiliations;
 
         private Builder() {
         }
@@ -32,18 +30,13 @@ public record TestNviContributor(String id,
             return this;
         }
 
-        public Builder withIsNviContributor(boolean isNviContributor) {
-            this.isNviContributor = isNviContributor;
-            return this;
-        }
-
-        public Builder withAffiliations(List<TestAffiliation> affiliations) {
+        public Builder withAffiliations(List<NviOrganization> affiliations) {
             this.affiliations = affiliations;
             return this;
         }
 
         public TestNviContributor build() {
-            return new TestNviContributor(id, isNviContributor, affiliations);
+            return new TestNviContributor(id, affiliations);
         }
     }
 }
